Engine repass + RecordFrame analyzer type

+++ Mutagen.Bethesda.Analyzers.Common/MutagenTopicBuilder.cs
@@ -5,7 +5,7 @@ namespace Mutagen.Bethesda.Analyzers
     public static class MutagenTopicBuilder
     {
         public static TopicDefinition FromDiscussion(
-            int id,
+            ushort id,
             string title,
             Severity severity)
         {
